Fix playback speed presets to match API's 0.1-increment validation
The preset list included 0.75/1.25/1.75, which aren't on the API's
0.1-step grid (0.5...3.0) and would always be rejected with a 400 if
selected. Also fixes a test assertion casting a JSON-decoded number to
Float (always nil — JSONSerialization produces Double/NSNumber) and
formats the player screen's speed label with fixed precision for
non-preset values, matching PlaybackSpeedOption.label.

+++ ios/Kuulla/KuullaTests/SettingsClientTests.swift
@@ -233,7 +233,9 @@ final class SettingsClientTests: MockedApiTestCase {
         XCTAssertTrue(requestedURL.absoluteString.hasSuffix("/api/settings/playback-speed"))
         XCTAssertEqual(capturedMethod, "PUT")
         let bodyJSON = try XCTUnwrap(JSONSerialization.jsonObject(with: XCTUnwrap(capturedBody)) as? [String: Any])
-        XCTAssertEqual(bodyJSON["playbackSpeed"] as? Float, 1.5)
+        // JSONSerialization decodes numeric values as NSNumber/Double, not Float — `as? Float`
+        // would fail to cast (nil), failing this assertion even though the body is correct.
+        XCTAssertEqual(bodyJSON["playbackSpeed"] as? Double, 1.5)
     }
 
     func testGetShowSettingsDecodesNullPlaybackSpeedOverrideAsNil() async throws {
